При повороте учитываь смещение Изменить градации зума

+++ app/src/main/java/coe/com/c0r0vans/GameObjects/Player.java
@@ -110,7 +110,7 @@ public class Player extends GameObject {
 
 
     @Override
-    public void changeMarkerSize(int Type) {
+    public void changeMarkerSize(float Type) {
         player.mark.setIcon(BitmapDescriptorFactory.fromResource(R.mipmap.marker));
     }
